+++ packages/ts-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @baunilha/ts-config
 
+## 1.0.3
+
+### Patch Changes
+
+- Add fixed build and initial native support
+
+## 1.0.2
+
+### Patch Changes
+
+- Add initial native components support
+
 ## 1.0.1
 
 ### Patch Changes